flattening the room and user icon loading into the same file
